ENH: simplify support options and fix a tree sup bug
1. add default support base pattern (lightning for tree support,
   rectilinear for normal)
2. change default support interface pattern to "Default" (concentric for
   absoluble material, rectilinear for others)
3. delete tree_support_with_infill, replace no-infill optio with "None" infill type
4. delete HAS_LIGHTNING_INFILL macro (lightning is impossible to be
   removed now)
5. fix the issue that tree support base may grow above interface by
   adding virtual nodes (distance_to_top<0).

(cherry picked from commit a2b41b1dfd9227e68ac6d971312133ec3382baa7)

+++ src/libslic3r/TreeSupport.hpp
@@ -223,7 +223,7 @@ public:
          , height(0.0)
         {}
 
-        Node(const Point position, const size_t distance_to_top, const bool skin_direction, const int support_roof_layers_below, const bool to_buildplate, Node* parent,
+        Node(const Point position, const int distance_to_top, const bool skin_direction, const int support_roof_layers_below, const bool to_buildplate, Node* parent,
              coordf_t     print_z_, coordf_t height_)
          : distance_to_top(distance_to_top)
          , position(position)
@@ -252,8 +252,9 @@ public:
 
         /*!
          * \brief The number of layers to go to the top of this branch.
+         * Negative value means it's a virtual node between support and overhang, which doesn't need to be extruded.
          */
-        size_t distance_to_top;
+        int distance_to_top;
 
         /*!
          * \brief The position of this node on the layer.
